refactor: keep service models and extends in api layer (#1655)
