+++ src/rubrix/server/apis/v0/handlers/token_classification_dataset_settings.py
@@ -9,11 +9,11 @@ from rubrix.server.apis.v0.models.dataset_settings import TokenClassificationSet
 from rubrix.server.apis.v0.validators.token_classification import DatasetValidator
 from rubrix.server.security import auth
 from rubrix.server.security.model import User
-from rubrix.server.services.datasets import DatasetsService, ServiceDatasetSettings
+from rubrix.server.services.datasets import DatasetsService, ServiceBaseDatasetSettings
 
-__svc_settings_class__: Type[ServiceDatasetSettings] = type(
+__svc_settings_class__: Type[ServiceBaseDatasetSettings] = type(
     f"{TaskType.token_classification}_DatasetSettings",
-    (ServiceDatasetSettings, TokenClassificationSettings),
+    (ServiceBaseDatasetSettings, TokenClassificationSettings),
     {},
 )
