--- src/rubrix/server/apis/v0/models/info.py
@@ -1,31 +1,0 @@
-#  coding=utf-8
-#  Copyright 2021-present, the Recognai S.L. team.
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
-
-from typing import Any, Dict
-
-from pydantic import BaseModel
-
-
-class ApiInfo(BaseModel):
-    """Basic api info"""
-
-    rubrix_version: str
-
-
-class ApiStatus(ApiInfo):
-    """The Rubrix api status model"""
-
-    elasticsearch: Dict[str, Any]
-    mem_info: Dict[str, Any]
