+++ src/rubrix/server/apis/v0/validators/text_classification.py
@@ -12,11 +12,11 @@ from rubrix.server.apis.v0.models.text_classification import (
 )
 from rubrix.server.errors import BadRequestError, EntityNotFoundError
 from rubrix.server.security.model import User
-from rubrix.server.services.datasets import DatasetsService, ServiceDatasetSettings
+from rubrix.server.services.datasets import DatasetsService, ServiceBaseDatasetSettings
 
-__svc_settings_class__: Type[ServiceDatasetSettings] = type(
+__svc_settings_class__: Type[ServiceBaseDatasetSettings] = type(
     f"{TaskType.text_classification}_DatasetSettings",
-    (ServiceDatasetSettings, TextClassificationSettings),
+    (ServiceBaseDatasetSettings, TextClassificationSettings),
     {},
 )
