+++ src/rubrix/server/apis/v0/validators/token_classification.py
@@ -13,11 +13,11 @@ from rubrix.server.apis.v0.models.token_classification import (
 )
 from rubrix.server.errors import BadRequestError, EntityNotFoundError
 from rubrix.server.security.model import User
-from rubrix.server.services.datasets import DatasetsService, ServiceDatasetSettings
+from rubrix.server.services.datasets import DatasetsService, ServiceBaseDatasetSettings
 
-__svc_settings_class__: Type[ServiceDatasetSettings] = type(
+__svc_settings_class__: Type[ServiceBaseDatasetSettings] = type(
     f"{TaskType.token_classification}_DatasetSettings",
-    (ServiceDatasetSettings, TokenClassificationSettings),
+    (ServiceBaseDatasetSettings, TokenClassificationSettings),
     {},
 )
