+++ src/rubrix/server/services/tasks/text2text/__init__.py
@@ -1,0 +1,1 @@
+from .service import Text2TextService
