+++ src/rubrix/server/services/tasks/text_classification/__init__.py
@@ -1,0 +1,2 @@
+from .labeling_rules_service import LabelingService
+from .service import TextClassificationService
