+++ src/rubrix/server/services/tasks/token_classification/__init__.py
@@ -1,0 +1,1 @@
+from .service import TokenClassificationService
