+++ tests/server/commons/test_records_dao.py
@@ -1,8 +1,8 @@
 import pytest
 
 from rubrix.server.apis.v0.models.commons.model import TaskType
-from rubrix.server.apis.v0.models.datasets import DatasetDB
 from rubrix.server.backend.elasticsearch import ElasticsearchBackend
+from rubrix.server.daos.models.datasets import BaseDatasetDB
 from rubrix.server.daos.records import DatasetRecordsDAO
 from rubrix.server.errors import MissingDatasetRecordsError
 
@@ -11,5 +11,7 @@ def test_raise_proper_error():
     dao = DatasetRecordsDAO.get_instance(ElasticsearchBackend.get_instance())
     with pytest.raises(MissingDatasetRecordsError):
         dao.search_records(
-            dataset=DatasetDB(name="mock-notfound", task=TaskType.text_classification)
+            dataset=BaseDatasetDB(
+                name="mock-notfound", task=TaskType.text_classification
+            )
         )
