+++ tests/server/datasets/test_dao.py
@@ -15,10 +15,10 @@
 
 import pytest
 
-from rubrix.server.apis.v0.models.datasets import DatasetDB
 from rubrix.server.backend.elasticsearch import create_es_wrapper
 from rubrix.server.commons.models import TaskType
 from rubrix.server.daos.datasets import DatasetsDAO
+from rubrix.server.daos.models.datasets import BaseDatasetDB
 from rubrix.server.daos.records import DatasetRecordsDAO
 from rubrix.server.errors import ClosedDatasetError
 
@@ -30,7 +30,7 @@ dao = DatasetsDAO.get_instance(es_wrapper, records)
 def test_retrieve_ownered_dataset_for_no_owner_user():
     dataset = "test_retrieve_owned_dataset_for_no_owner_user"
     created = dao.create_dataset(
-        DatasetDB(name=dataset, owner="other", task=TaskType.text_classification),
+        BaseDatasetDB(name=dataset, owner="other", task=TaskType.text_classification),
     )
     assert dao.find_by_name(created.name, owner=created.owner) == created
     assert dao.find_by_name(created.name, owner=None) == created
@@ -41,7 +41,7 @@ def test_close_dataset():
     dataset = "test_close_dataset"
 
     created = dao.create_dataset(
-        DatasetDB(name=dataset, owner="other", task=TaskType.text_classification),
+        BaseDatasetDB(name=dataset, owner="other", task=TaskType.text_classification),
     )
 
     dao.close(created)
