+++ tests/server/token_classification/test_model.py
@@ -20,12 +20,12 @@ from rubrix._constants import MAX_KEYWORD_LENGTH
 from rubrix.server.apis.v0.models.commons.model import PredictionStatus
 from rubrix.server.apis.v0.models.token_classification import (
     CreationTokenClassificationRecord,
-    EntitySpan,
     TokenClassificationAnnotation,
     TokenClassificationQuery,
     TokenClassificationRecord,
 )
 from rubrix.server.backend.search.query_builder import EsQueryBuilder
+from rubrix.server.services.tasks.token_classification.model import EntitySpan
 
 
 def test_char_position():
